\name{microscopy}
\alias{microscopy}
\docType{data}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Data of uncut fiber lengths in the increment core}

\description{Simulated data of lengths of uncut fibers in the increment core (as measured by microscopy), under the assumption that the true lengths of those fibers that at least partially appear in the increment core follow a generalized gamma distribution with parameters \code{par=c(2.4, 3.3, 1.5)}.
 Useful for illustrating the use of \code{fled()}.
}
\usage{
data(microscopy)
}
%- maybe also `usage' for other objects documented here.

\format{ \code{microscopy} is a vector of 300 fiber lengths simulated from the density 

\code{dx.fibers.micro(x, par, r=2.5, model="ggamma")} with parameters \code{par=c(2.4, 3.3, 1.5)} 
} 


%\details{The data ...
%}

%\source{
%The data are adapted from ...., where the original source is given as: 
%}

%\examples{
%}








