\name{print.fled}
\alias{print.fled}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Print a fiber length determination object.}
\description{ The default print method for a \code{fled} object. 
}
\usage{
\method{print}{fled}(x, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{ 
\item{x, ...}{ fitted model objects of class \code{fled} as produced by \code{fled()}.}
}
\details{ Prints the model, type of data, estimated model parameters, optimized value of the minus log likelihood and number of observations supplied. 

}


\author{
Sara \enc{Sjöstedt}{Sjostedt} de Luna, Konrad Abramowicz, Natalya Pya Arnqvist
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%    \code{\link{S2.test}}, \code{}    
%}

%\examples{

%}











